% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_utils.R
\name{build_measure}
\alias{build_measure}
\title{Build typical measurement function using split/measure/pool approach.}
\usage{
build_measure(token, token_measure, token_pool)
}
\arguments{
\item{token}{The token to split (default: words). See `tidytext::unnest_tokens`
for more options.}

\item{token_measure}{A token measurement function. See `measure_tokens`.}

\item{token_pool}{A token measurement pooling function. See `pool_measured_tokens`.}
}
\value{
A function which returns a numeric measurement vector given a text
response.
}
\description{
Build typical measurement function using split/measure/pool approach.
}
