% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{compute_p_val}
\alias{compute_p_val}
\title{Compute p-value based on permutation test statistics.}
\usage{
compute_p_val(stat_observed, stats_permuted)
}
\description{
Compute p-value based on permutation test statistics.
}
