% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_word2vec_targets.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Compute cosine similarity between `token` and tokens in `target`.}
\usage{
cosine_similarity(token, targets, model)
}
\arguments{
\item{token}{A character string.}

\item{targets}{A vector of relevant target tokens.}
}
\value{
A vector of cosine similarities.
}
\description{
Compute cosine similarity between `token` and tokens in `target`.
}
