% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest}
\alias{digest}
\title{Digest response measurements into observed statistic, permuted statistics,
and a p-value.}
\usage{
digest(data, test_formula, stat_compute, num_permutations = 1000)
}
\arguments{
\item{data}{A data frame with columns `arm` and `measurement`. The column
`arm` should be a character column, specifying which treatment arm the
particular observation is in. The column `measurement` should be a list
column, and each measurement should be a numeric vector.}

\item{test_formula}{A formula specifying treatment arms to compare. The left-hand
side is considered the treatment, the right-hand side the control.}

\item{stat_compute}{A function which accepts a data frame (with the same
specifications as those of `data`) and returns an arbitrary _numeric_ test
statistic.}

\item{num_permutations}{The number of permutations of the treatment arm to
perform for the permutation test.}
}
\value{
A data frame containing the observed and permuted statistics, along with
their associated p-value.
}
\description{
Digest response measurements into observed statistic, permuted statistics,
and a p-value.
}
