% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_plot.R
\name{digest_plot}
\alias{digest_plot}
\title{Plot permutation test data returned by `digest`.}
\usage{
digest_plot(data)
}
\arguments{
\item{data}{A data frame, as returned by `digest`. At a minimum, this data
frame should have columns `test`, `stat_observed`, `stats_permuted`, and `p_val`.}
}
\value{
A ggplot plot object.
}
\description{
Plot permutation test data returned by `digest`.
}
