% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Measure response text as a numeric vector.}
\usage{
measure(data, cols, measure_response, ...)
}
\arguments{
\item{data}{An arbitrary data frame.}

\item{cols}{A list of columns in `data` (must be character columns).}

\item{measure_response}{A function which accepts a vector of strings and returns
a vector of numeric vectors (each its own text measurement).}
}
\description{
Measure response text as a numeric vector.
}
