% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_count_targets.R
\name{measure_count_targets}
\alias{measure_count_targets}
\title{Measure free text response as a pooled set of target token counts.}
\usage{
measure_count_targets(response, targets)
}
\description{
Measure free text response as a pooled set of target token counts.
}
