% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_utils.R
\name{measure_tokens}
\alias{measure_tokens}
\title{Measure tokens using the given function.}
\usage{
measure_tokens(tokens, token_measure, ...)
}
\arguments{
\item{tokens}{A tibble of tokens in the format returned by `split_tokens`.}

\item{measure_token}{A function which accepts a token (along with
additional parameters in `...`) and returns a numeric vector.}
}
\value{
A tibble, as returned by `split_tokens`, along with a new column
`measurement`, as returned by `measure_token_fn` for each token.
}
\description{
Measure tokens using the given function.
}
