% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_word2vec_targets.R
\name{measure_word2vec_targets}
\alias{measure_word2vec_targets}
\title{Measure free text response as a pooled set of cosine similarity between
response words and target words.}
\usage{
measure_word2vec_targets(response, targets, model)
}
\arguments{
\item{response}{String to measure.}

\item{targets}{Target words to consider in measuring `response`.}

\item{model}{Word2Vec model, which maps words to numeric vectors.}
}
\description{
Measure free text response as a pooled set of cosine similarity between
response words and target words.
}
