% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_utils.R
\name{pmax_list}
\alias{pmax_list}
\title{Take element-wise maximum of list of vectors.}
\usage{
pmax_list(vectors)
}
\arguments{
\item{vectors}{A list of numeric vectors.}
}
\value{
The element-wise maximum of all input vectors.
}
\description{
This is a common function for pooling token measurements.
}
