% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_utils.R
\name{pool_measured_tokens}
\alias{pool_measured_tokens}
\title{Pool token measurements into a single response measurement.}
\usage{
pool_measured_tokens(tokens, token_pool)
}
\arguments{
\item{tokens}{A tibble of tokens and measurements as returned by `measure_tokens`.}

\item{token_pool}{A function which accepts a list of token measurements
and returns a numeric vector representing a single response measurement.}
}
\value{
A single pooled response measurement.
}
\description{
Pool token measurements into a single response measurement.
}
