% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{set_arms}
\alias{set_arms}
\title{Filters data down to treatment arms specified in `test_formula` and mutates `arm`
to be 1 if an observation is in the treatment group and 0 otherwise.}
\usage{
set_arms(data, test_formula)
}
\description{
Filters data down to treatment arms specified in `test_formula` and mutates `arm`
to be 1 if an observation is in the treatment group and 0 otherwise.
}
