% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{set_trmt}
\alias{set_trmt}
\title{Filters data down to treatment arms specified in `test_formula` and sets
column `trmt` to a logical, specifying if the observation is from the treatment
or control group.}
\usage{
set_trmt(data, test_formula)
}
\description{
Filters data down to treatment arms specified in `test_formula` and sets
column `trmt` to a logical, specifying if the observation is from the treatment
or control group.
}
