% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_utils.R
\name{spread_measurements}
\alias{spread_measurements}
\title{Make each element in measurement vectors a new column.}
\usage{
spread_measurements(data)
}
\arguments{
\item{data}{A data frame with list column `measurement`.}
}
\value{
A new data frame, with measurement vectors spread across multiple
columns.
}
\description{
Make each element in measurement vectors a new column.
}
