% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summed_mean_difference.R
\name{stat_summed_mean_difference}
\alias{stat_summed_mean_difference}
\title{Mean difference test statistic.}
\usage{
stat_summed_mean_difference(data)
}
\arguments{
\item{data}{A data frame which contains columns `trmt` and `measurement`.}
}
\description{
Mean difference test statistic.
}
