% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_utils.R
\name{sum_list}
\alias{sum_list}
\title{Sum list of numeric vectors.}
\usage{
sum_list(vectors)
}
\value{
The sum of all input vectors.
}
\description{
This is a common function for pooling token measurements.
}
