% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_count_targets.R
\name{binary_count}
\alias{binary_count}
\title{Check equality between `token` and tokens in `target`.}
\usage{
binary_count(token, targets)
}
\arguments{
\item{token}{A character string.}

\item{targets}{A vector of relevant target tokens.}
}
\value{
A vector of counts (a one-hot encoding).
}
\description{
Check equality between `token` and tokens in `target`.
}
