% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_plot.R
\name{digest_plot}
\alias{digest_plot}
\title{Plot permutation test data returned by `digest`.}
\usage{
digest_plot(data, binwidth = 0.05)
}
\arguments{
\item{data}{A data frame, as returned by `digest`. At a minimum, this data
frame should have columns `test`, `stat_observed`, `stats_permuted`, and `p_val`.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that calculates width from unscaled x. Here, "unscaled x"
refers to the original x values in the data, before application of any
scale transformation. When specifying a function along with a grouping
structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}
}
\value{
A ggplot plot object.
}
\description{
Plot permutation test data returned by `digest`.
}
