% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_utils.R
\name{split_tokens}
\alias{split_tokens}
\title{Split response into tokens.}
\usage{
split_tokens(response, ...)
}
\arguments{
\item{response}{A vector of character strings.}

\item{...}{Additional parameters passed to `tidytext::unnest_tokens`.}
}
\value{
A tibble with a column of tokens parsed from `response`.
}
\description{
Most of the work is done by `unnest_tokens` from the tidytext package.
By default, this splits the response into words.
}
