% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_utils.R
\name{split_tokens}
\alias{split_tokens}
\title{Split each response into tokens.}
\usage{
split_tokens(responses, ...)
}
\arguments{
\item{responses}{A vector of character strings.}
}
\value{
A tibble with a column indicating the index of the response from
which the token originates, along with the token itself.
}
\description{
Most of the work is done by `unnest_tokens` from the tidytext package.
By default, this splits responses into words.
}
