% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_classifier.R
\name{stat_classifier}
\alias{stat_classifier}
\title{Classifier test statistic which returns the concordance (C-statistic, AUC, etc.)
of a classifier trained on measurements to predict treatment arm.}
\usage{
stat_classifier(data)
}
\arguments{
\item{data}{A data frame which contains columns `trmt` and `measurement`.}
}
\description{
Classifier test statistic which returns the concordance (C-statistic, AUC, etc.)
of a classifier trained on measurements to predict treatment arm.
}
